+++ Tests/run-document-layout.php
@@ -14,7 +14,7 @@ require FS_FOLDER . '/vendor/autoload.php';
 require FS_FOLDER . '/config.php';
 \FacturaScripts\Core\Kernel::init();
 
-putenv('BEPLY_PDF_PRECISE_BOTTOM_ANCHOR=');
+putenv('BEPLY_PDF_PRECISE_BOTTOM_ANCHOR=0');
 
 use FacturaScripts\Core\Tools;
 use FacturaScripts\Plugins\BeplyPDFStudio\Lib\BeplyPdfConfig;